feedback: add an abstraction of a Stream<State> -> Stream<Mutation>
Basically, a Feedback is simply a function transforming an incoming State into a Mutation while
eventually performing side effects.

There are many ways of creating a feedback:
- in its purest form it is a Stream<State> -> Stream<Mutation> function
- it can also be a State -> Stream<Mutation> function to ease its usage
- ...

A feedback can be composed of several feedbacks. That is why there are many initializers
to the Feedback protocol.

Function builders are leveraged to be able to instantiate a Feedback system in a declarative way.